requirements.txt
streamlit==1.31.0
opencv-python==4.9.0.80
ultralytics==8.1.0
numpy==1.26.3
Pillow==10.2.0
